requirements.txt
astroid==2.2.5
branca==0.3.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
Flask==1.0.2
folium==0.8.3
future==0.17.1
geocoder==1.38.1
gunicorn==19.9.0
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.3
pandas==0.24.2
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
ratelim==0.1.6
requests==2.21.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.11.1
